+++ youtube-atlas-backend/src/main/java/com/yongsoo/youtubeatlasbackend/game/api/CreateScheduledSellOrderRequest.java
@@ -9,7 +9,7 @@ import jakarta.validation.constraints.NotNull;
 public record CreateScheduledSellOrderRequest(
     @NotNull Long positionId,
     @NotNull @Min(1) @Max(200) Integer targetRank,
-    @NotNull @Min(1) Integer quantity,
+    @NotNull @Min(1) Long quantity,
     ScheduledSellTriggerDirection triggerDirection
 ) {
 }
