+++ youtube-atlas-backend/src/main/java/com/yongsoo/youtubeatlasbackend/game/api/LeaderboardEntryResponse.java
@@ -18,7 +18,7 @@ public record LeaderboardEntryResponse(
     Double profitRatePercent,
     Long realizedPnlPoints,
     Long unrealizedPnlPoints,
-    Integer openPositionCount,
+    Long openPositionCount,
     Boolean me
 ) {
 }
